[update] データ取得とフロントのAPI繋ぎ込み
