+++ front/src/hooks/pages/useLogin.ts
@@ -7,6 +7,7 @@ import { SubmitHandler, useForm } from 'react-hook-form'
 import { app } from '~/libs/firebase'
 import yup from '~/libs/yup/locale'
 import { isFirebaseError } from '~/types/firebase'
+import { setLocalStorage} from "~/utils/localstorage";
 
 export type LoginFormType = {
   email: string
@@ -52,12 +53,15 @@ export const useLogin = () => {
       setStatus('loading')
 
       try {
-        await signInWithEmailAndPassword(auth, email, password)
+        const {user: {uid: userId}} = await signInWithEmailAndPassword(auth, email, password)
         const token = await auth.currentUser?.getIdToken(true)
-        localStorage.setItem('token', token ?? '')
+
+        if(!token) return
+        await setLocalStorage('token', token)
         await push({
-          pathname: '/teams',
+          pathname: `/${userId}`,
         })
+
       } catch (error) {
         if (!(error instanceof Error && isFirebaseError(error))) return
