+++ front/src/libs/apolloClient/index.ts
@@ -43,6 +43,7 @@ function createApolloClient() {
                 },
             },
         }),
+        connectToDevTools: true
     })
 }
