+++ front/src/utils/isString.ts
@@ -1,0 +1,3 @@
+export const isString = (value: unknown): value is string => {
+    return typeof value === "string"
+}
